requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
backoff==2.2.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
dataclasses-json==0.6.7
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.110.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.14
langchain-core==0.3.63
langchain-google-genai==0.0.11
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.0.22
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.43
lxml==5.4.0
marshmallow==3.26.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pillow==10.2.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic-settings==2.2.1
pydantic==2.6.4
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==5.5.0
pytesseract==0.3.10
pytest==8.3.5
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.6
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
unstructured==0.12.2
unstructured-client==0.36.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.29.0
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
